vm: import construction → respective VM impls
I always found it weird that we had VM specific code in what's a generic
part of the near-vm-runner. Well -- it has actually started bothering me
in real ways, so it gets moved.

I didn't do very much to make sure it ends up pretty. And I don't think
I want to for the VMs that're not wasmtime or near_vm. Good thing is
that the wasmer0/2 code can be largely ignored sans changes to `VMLogic`
(which I'm considering addressing next...)
